Move mock constants into own modue
