+++ p2panda-rs/src/test_utils/mocks/materialisation/processor.rs
@@ -1,8 +1,8 @@
 // SPDX-License-Identifier: AGPL-3.0-or-later
 
+use crate::test_utils::mocks::constants::PERMISSIONS_SCHEMA_HASH;
 use crate::test_utils::mocks::logs::LogEntry;
 use crate::test_utils::mocks::materialisation::Materialiser;
-use crate::test_utils::mocks::utils::PERMISSIONS_SCHEMA_HASH;
 
 /// Get all permission entries from this author
 pub fn author_permission_entries(entries: &[LogEntry], author: &str) -> Vec<LogEntry> {
