+++ p2panda-rs/src/test_utils/mocks/mod.rs
@@ -3,6 +3,7 @@
 //! Mock p2panda node and client for demo and testing purposes.
 
 mod client;
+pub mod constants;
 pub mod logs;
 pub mod materialisation;
 mod node;
